chore/client: refactor Reward page
