+++ client/src/pages/Rewards/components/DepositInfo/index.scss
@@ -1,10 +1,4 @@
 .deposit-info {
-    padding: 20px;
-    border-radius: 20px;
-    background-color: var(--primary);
-    display: flex;
-    flex-direction: column;
-
     &__row {
         margin-bottom: 20px;
         &:last-child {
