+++ client/src/pages/Rewards/components/TransactionStatus/index.scss
@@ -1,0 +1,20 @@
+.transaction-status {
+    &__status-row {
+        margin-bottom: 10px;
+        display: flex;
+        flex-direction: row;
+        &:last-child {
+            margin-bottom: 0;
+        }
+        & > div {
+            margin-left: 5px;
+            &:first-child {
+                margin-left: 0;
+            }
+        }
+        a {
+            cursor: pointer;
+            transition: 0.1s;
+        }
+    }
+}
